WIP elevator ui hierarchy
