+++ randovania/games/common/gui/elevators_tab_common.py
@@ -7,19 +7,14 @@ from typing import TYPE_CHECKING
 from PySide6 import QtCore, QtWidgets
 
 from randovania.game_description.db.dock_node import DockNode
-from randovania.games.prime2.exporter.patch_data_factory import should_keep_elevator_sounds
-from randovania.games.prime2.layout.echoes_configuration import EchoesConfiguration
-from randovania.gui.generated.preset_elevators_ui import Ui_PresetElevators
 from randovania.gui.lib import signal_handling
 from randovania.gui.lib.node_list_helper import NodeListHelper
 from randovania.gui.preset_settings.preset_tab import PresetTab
 from randovania.layout.lib.teleporters import (
-    TeleporterConfiguration,
     TeleporterList,
     TeleporterShuffleMode,
     TeleporterTargetList,
 )
-from randovania.lib import enum_lib
 from randovania.patching.prime import elevators
 
 if TYPE_CHECKING:
@@ -27,12 +22,15 @@ if TYPE_CHECKING:
     from randovania.game_description.db.area_identifier import AreaIdentifier
     from randovania.game_description.db.node_identifier import NodeIdentifier
     from randovania.game_description.game_description import GameDescription
+    from randovania.games.common.prime_family.layout.lib.prime_triology_teleporters import (
+        PrimeTriologyTeleporterConfiguration,
+    )
     from randovania.gui.lib.window_manager import WindowManager
     from randovania.interface_common.preset_editor import PresetEditor
     from randovania.layout.preset import Preset
 
 
-class PresetElevators(PresetTab, Ui_PresetElevators, NodeListHelper):
+class PresetElevatorsCommon(PresetTab, NodeListHelper):
     _elevator_source_for_location: dict[NodeIdentifier, QtWidgets.QCheckBox]
     _elevator_source_destination: dict[NodeIdentifier, NodeIdentifier | None]
     _elevator_target_for_region: dict[str, QtWidgets.QCheckBox]
@@ -59,8 +57,6 @@ class PresetElevators(PresetTab, Ui_PresetElevators, NodeListHelper):
         self.elevators_description_label.setText("".join(descriptions))
 
         self.elevators_combo.currentIndexChanged.connect(self._update_elevator_mode)
-        signal_handling.on_checked(self.skip_final_bosses_check, self._update_require_final_bosses)
-        signal_handling.on_checked(self.elevators_allow_unvisited_names_check, self._update_allow_unvisited_names)
 
         # Elevator Source
         self._create_source_elevators()
@@ -187,7 +183,7 @@ class PresetElevators(PresetTab, Ui_PresetElevators, NodeListHelper):
 
     def on_preset_changed(self, preset: Preset):
         config = preset.configuration
-        config_elevators: TeleporterConfiguration = config.elevators
+        config_elevators: PrimeTriologyTeleporterConfiguration = config.elevators
 
         descriptions = [
             "<p>Controls where each elevator connects to.</p>",
@@ -195,11 +191,6 @@ class PresetElevators(PresetTab, Ui_PresetElevators, NodeListHelper):
             f' {config_elevators.mode.description}</p>',
         ]
         self.elevators_description_label.setText("".join(descriptions))
-
-        sound_bug_warning = False
-        if isinstance(config, EchoesConfiguration):
-            sound_bug_warning = not should_keep_elevator_sounds(config)
-        self.elevators_help_sound_bug_label.setVisible(sound_bug_warning)
 
         signal_handling.set_combo_with_value(self.elevators_combo, config_elevators.mode)
         can_shuffle_source = config_elevators.mode not in (TeleporterShuffleMode.VANILLA,
@@ -241,92 +232,10 @@ class PresetElevators(PresetTab, Ui_PresetElevators, NodeListHelper):
         self.elevators_source_group.setVisible(can_shuffle_source)
         self.elevators_target_group.setVisible(config_elevators.has_shuffled_target)
         self.elevators_target_group.setEnabled(config_elevators.has_shuffled_target)
-        self.skip_final_bosses_check.setChecked(config_elevators.skip_final_bosses)
-        self.elevators_allow_unvisited_names_check.setChecked(config_elevators.allow_unvisited_room_names)
         self.update_node_list(
             config_elevators.excluded_targets.locations,
             True,
             self._elevator_target_for_region,
             self._elevator_target_for_area,
             self._elevator_target_for_node
-        )
-
-
-class PresetElevatorsPrime1(PresetElevators):
-    compatible_modes = [
-        value for value in enum_lib.iterate_enum(TeleporterShuffleMode)
-        if value != TeleporterShuffleMode.ECHOES_SHUFFLED
-    ]
-
-    def __init__(self, editor: PresetEditor, game_description: GameDescription, window_manager: WindowManager):
-        super().__init__(editor, game_description, window_manager)
-        self.elevators_allow_unvisited_names_check.setVisible(False)
-        self.elevators_line_3.setVisible(False)
-        self.elevators_help_list_label.setVisible(False)
-
-        self.skip_final_bosses_check.setText("Go directly to credits from Artifact Temple")
-        self.skip_final_bosses_label.setText("""<html><head/><body>
-        <p>Change the teleport in Artifact Temple to go directly to the credits, skipping the final bosses.</p>
-        <p>This changes the requirements to <span style=" font-weight:600;">not need the final bosses</span>,
-        turning certain items optional such as Plasma Beam.</p></body></html>
-        """)
-
-
-class PresetElevatorsPrime2(PresetElevators):
-    compatible_modes = list(enum_lib.iterate_enum(TeleporterShuffleMode))
-
-    custom_weights = {
-            "Great Temple": 0,
-            "Agon Wastes": 1,
-            "Torvus Bog": 2,
-            "Sanctuary Fortress": 3,
-            "Temple Grounds": 5,
-        }
-    def __init__(self, editor: PresetEditor, game_description: GameDescription, window_manager: WindowManager):
-        super().__init__(editor, game_description, window_manager)
-
-
-class PresetElevatorsPrime3(PresetElevators):
-    compatible_modes = [
-        value for value in enum_lib.iterate_enum(TeleporterShuffleMode)
-        if value != TeleporterShuffleMode.ECHOES_SHUFFLED
-    ]
-
-    def __init__(self, editor: PresetEditor, game_description: GameDescription, window_manager: WindowManager):
-        super().__init__(editor, game_description, window_manager)
-
-        self.elevators_allow_unvisited_names_check.setVisible(False)
-        self.elevators_line_3.setVisible(False)
-        self.elevators_help_list_label.setVisible(False)
-
-        self.skip_final_bosses_check.setVisible(False)
-        self.skip_final_bosses_label.setVisible(False)
-
-        self.elevators_description_label.setText(
-            self.elevators_description_label.text().replace("elevator", "teleporter")
-        )
-
-class PresetElevatorsDread(PresetElevators):
-    compatible_modes = [
-        TeleporterShuffleMode.VANILLA,
-        TeleporterShuffleMode.TWO_WAY_RANDOMIZED,
-        TeleporterShuffleMode.TWO_WAY_UNCHECKED,
-        TeleporterShuffleMode.ONE_WAY_ELEVATOR,
-        TeleporterShuffleMode.ONE_WAY_ELEVATOR_REPLACEMENT,
-    ]
-
-    def __init__(self, editor: PresetEditor, game_description: GameDescription, window_manager: WindowManager):
-        super().__init__(editor, game_description, window_manager)
-
-        self.elevators_allow_unvisited_names_check.setVisible(False)
-        self.elevators_line_3.setVisible(False)
-        self.elevators_help_list_label.setVisible(False)
-
-        self.skip_final_bosses_check.setVisible(False)
-        self.skip_final_bosses_label.setVisible(False)
-
-        self.elevators_target_group.setVisible(False)
-
-        self.elevators_description_label.setText(
-            self.elevators_description_label.text().replace("elevator", "teleporter")
         )
